# swagger.json
{
  "openapi": "3.0.0",
  "components": {
    "securitySchemes": {
      "bearerAuth" : {  
        "type" : "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security":{
    "bearerAuth" : []
  },
  "info": {
    "title": "Proyecto Flutter",
    "description": "API para registrar, listar, y editar Barberos, Clientes, Usuarios y Servicios",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": " http://127.0.0.1:4000"
    },
    {
      "url": "http://10.2.125.234:4000"
    }
  ],
...
  "paths": {
...
    "/obtenerBarberos": {
      "get": {
        "security":[{
          "bearerAuth" : []
        }],
        "tags":["Barbero"],
        "summary": "Buscar barbero registrados en el sistema",
        "description":"Este endpoint permite obtener un listado completo de los barberos registrados en el sistema.",
        "responses": {
          "200": {
            "description": "Listado de barberos",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                  "bar_usu_id"    : { "type": "integer" },
                  "usu_nombre"    : { "type": "string" },
                  "usu_apellido"  : { "type": "string" },
                  "usu_telefono"  : { "type": "string" },
                  "usu_correo"    : { "type": "string" },
                  "usu_tipo_doc"  : { "type": "string" },
                  "usu_num_doc"   : { "type": "string" }, 
                  "usu_usuario"   : { "type": "string" },
                  "usu_contraseña": { "type": "string" },
                  "usu_estado"    : { "type": "string" },
                  "usu_genero"    : { "type": "string" },
                  "usu_salario"   : { "type": "string" }
                  },
                  "example" : {
                    "bar_salario": "1500000",
                    "bar_usu_id": 3,
                    "usu_apellido": "Rodríguez",
                    "usu_contraseña": "claveAndres789",
                    "usu_correo": "andres.rodriguez@example.com",
                    "usu_estado": "activo",
                    "usu_genero": "Masculino",
                    "usu_nombre": "Andrés",
                    "usu_num_doc": "9876543210",
                    "usu_telefono": "3201122334",
                    "usu_tipo_doc": "ce",
                    "usu_usuario": "andresr"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Listado de barberos",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "mensaje": { "type" : "string" }
                  },
                  "example" : {
                    "mensaje": "Ningun barbero obtenido."
                  }
                }
              }
            }
          }
        }
      }
    },
    "/registrarBarbero": {
      "post": {
        "security":[{
          "bearerAuth" : []
        }],
        "tags":["Barbero"],
        "summary": "Registar un nuevo barbero",
        "description": "Este endpoint permite registrar un nuevo barbero en el sistema. Requiere enviar un objeto JSON con los siguentes datos como el bar_salario y el bar_usu_id.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["'bar_salario', 'bar_usu_id'"],
                "properties": {
                  "bar_salario" : { "type": "string" }, 
                  "bar_usu_id"  : { "type": "integer" }
                },
                "example" : {
                  "bar_salario": "1500000",
                  "bar_usu_id": 3
                } 
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Devuelve un mensaje indicando que el usuario fue registrado correctamente.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "properties": {
                    "bar_salario" : { "type": "string" }, 
                    "bar_usu_id"  : { "type": "integer" }
                  },
                  "example" : {
                    "bar_salario": "1500000",
                    "bar_usu_id": 3
                  }
                }
              }
            }
          },
          "404": {
            "description": "error de usuario inexistente.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "mensaje": { "type": "string" }
                  },
                  "example": {
                    "mensaje": "No existe un usuario registrado con ese ID"
                  }
                }
              }
            }
          } ,  
          "409": {
            "description": "error de entrada duplicada.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "mensaje": { "type": "string" }
                  },
                  "example": {
                    "mensaje": "Ya existe un Barbero registrado con ese ID"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/obtenerDetalleServicios": {
      "get": {
        "security":[{
          "bearerAuth" : []
        }],
        "tags": ["DetalleServicio"],
        "summary": "Listar todos los detalles de servicio",
        "description": "Este endpoint permite obtener un listado de los detalles de servicios realizados en el sistema.",
        "responses": {
          "200": {
            "description": "Respuesta exitosa que contiene la lista de detalles de servicio",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "dtll_id":       { "type": "integer" },
                      "serv_tipo":     { "type": "string" },
                      "serv_precio":   { "type": "string" },
                      "cli_nombre":    { "type": "string" },
                      "cli_apellido":  { "type": "string" },
                      "bar_nombre":    { "type": "string" },
                      "bar_apellido":  { "type": "string" }
                    },
                    "example": {
                      "dtll_id": 5,
                      "serv_tipo": "corte + barba",
                      "serv_precio": "20000",
                      "cli_nombre": "Carlos",
                      "cli_apellido": "Pérez",
                      "bar_nombre": "Luis",
                      "bar_apellido": "Gómez"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "No se encontraron detalles de servicio.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "mensaje": { "type": "string" }
                  },
                  "example": {
                    "mensaje": "Ningún Detalle de Servicio Obtenido"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/registrarDetalleServicio": {
      "post": {
        "security":[{
          "bearerAuth" : []
        }],
        "tags": ["DetalleServicio"],
        "summary": "Registrar un nuevo detalle de servicio",
        "description": "Este endpoint permite registrar un nuevo detalle de servicio. Requiere enviar un objeto JSON con los campos dtll_serv_id, dtll_cli_id y dtll_bar_id.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["dtll_serv_id", "dtll_cli_id", "dtll_bar_id"],
                "properties": {
                  "dtll_serv_id": { "type": "integer" },
                  "dtll_cli_id":  { "type": "integer" },
                  "dtll_bar_id":  { "type": "integer" }
                },
                "example": {
                  "dtll_serv_id": 2,
                  "dtll_cli_id": 4,
                  "dtll_bar_id": 1
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Detalle registrado exitosamente.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "mensaje": { "type": "string" }
                  },
                  "example": {
                    "mensaje": "Se ha registrado el Detalle del Servicio Realizado"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Campos faltantes o vacíos.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "mensaje": { "type": "string" }
                  },
                  "example": {
                    "mensaje": "Faltan campos por rellenar"
                  }
                }
              }
            }
          },
          "404": {
            "description": "ID de servicio, cliente o barbero no encontrado.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "mensaje": { "type": "string" }
                  },
                  "example": {
                    "mensaje": "Uy, parece que no hay ningún servicio con ese ID"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/editarDetalleServicio/{dtll_id}": {
      "put": {
        "security":[{
          "bearerAuth" : []
        }],
        "tags": ["DetalleServicio"],
        "summary": "Editar un detalle de servicio existente",
        "description": "Este endpoint permite editar un detalle de servicio existente mediante su ID. Requiere un objeto JSON con los campos dtll_serv_id, dtll_cli_id y dtll_bar_id.",
        "parameters": [
          {
            "name": "dtll_id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["dtll_serv_id", "dtll_cli_id", "dtll_bar_id"],
                "properties": {
                  "dtll_serv_id": { "type": "integer" },
                  "dtll_cli_id":  { "type": "integer" },
                  "dtll_bar_id":  { "type": "integer" }
                },
                "example": {
                  "dtll_serv_id": 3,
                  "dtll_cli_id": 5,
                  "dtll_bar_id": 2
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Detalle de servicio editado exitosamente.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "mensaje": { "type": "string" }
                  },
                  "example": {
                    "mensaje": "Se ha editado el Detalle del Servicio Realizado"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Campos faltantes o vacíos.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "mensaje": { "type": "string" }
                  },
                  "example": {
                    "mensaje": "Faltan campos por rellenar"
                  }
                }
              }
            }
          },
          "404": {
            "description": "ID no encontrado.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "mensaje": { "type": "string" }
                  },
                  "example": {
                    "mensaje": "No existe un detalle de servicio con ese ID"
                  }
                }
              }
            }
          }
        }
      }
    }
  }
}